--- a/vllm/attention/ops/triton_merge_attn_states.py
+++ b/vllm/attention/ops/triton_merge_attn_states.py
@@ -20,20 +20,6 @@ def merge_attn_states(
     num_query_heads = output.shape[1]
     head_size = output.shape[2]
     padded_head_size = triton.next_power_of_2(head_size)
-    # FA2 and FA3 have different behavior for when the sum-exp is 0, this namely
-    # arises with 0 len seqlens. FA3 returns -inf here while FA2 returns inf.
-    # If we see an inf assume FA2 and convert inf to -inf for consistency
-    # and correctness. Inf generally doesn't make sense in this context outside
-    # of undefined-behavior/FA2-case, so I think this a safe assumption.
-    # p_lse = float('-inf') if p_lse == float('inf') else p_lse
-    # s_lse = float('-inf') if s_lse == float('inf') else s_lse
-    # FIXME(DefTruth): To avoid affecting the generation of the Triton
-    # kernel, I chose to handle this processing outside the kernel. Otherwise,
-    # dynamically checking each value for infinity within the kernel might
-    # impact the execution efficiency of the generated Triton kernel.
-    prefix_lse[prefix_lse == torch.inf] = (-torch.inf)
-    suffix_lse[suffix_lse == torch.inf] = (-torch.inf)
-
     # TODO(woosuk): Use CUDA kernel instead of Triton to minimize CPU overhead.
     merge_attn_states_kernel[(num_tokens, num_query_heads)](
         output,
@@ -67,6 +53,18 @@ def merge_attn_states_kernel(
 
     p_lse = tl.load(prefix_lse + head_idx * num_tokens + token_idx)
     s_lse = tl.load(suffix_lse + head_idx * num_tokens + token_idx)
+    # FA2 and FA3 have different behavior for when the sum-exp is 0, this namely
+    # arises with 0 len seqlens. FA3 returns -inf here while FA2 returns inf.
+    # If we see an inf assume FA2 and convert inf to -inf for consistency
+    # and correctness. Inf generally doesn't make sense in this context outside
+    # of undefined-behavior/FA2-case, so I think this a safe assumption.
+    # FIXME(DefTruth): dynamically checking each value for infinity within 
+    # the kernel might impact the execution efficiency of the generated Triton 
+    # kernel (L20, etc.). To avoid affecting the generation of the Triton
+    # kernel, I have try to handle this processing outside the kernel, but
+    # still can't fix the performance issue. 
+    # p_lse = float('-inf') if p_lse == float('inf') else p_lse
+    # s_lse = float('-inf') if s_lse == float('inf') else s_lse
 
     max_lse = tl.maximum(p_lse, s_lse)
     p_lse = p_lse - max_lse