@DefTruth 2025-03-28
convert to a draft, seems still not really fix this performance problem.