@DefTruth 2025-03-28
I still don't know why these two lines will downgrade the performance in my test, @LucasWilkinson 
```python
p_lse = float('-inf') if p_lse == float('inf') else p_lse
 s_lse = float('-inf') if s_lse == float('inf') else s_lse
```