@LucasWilkinson 2025-03-28
Weird, good find though! I assumed this should translate into only a couple instructions i.e. a cmp and predicated mov but I guess not

Another option is to hack up FA2 but I think there should be a way to make this efficient.

Sorry on vacation till Tuesday afternoon, but can help dig in then if you don't make progress.