@DefTruth 2025-03-28
> #15492

may be we can try tl.inline_asm_elementwise (https://triton-lang.org/main/python-api/generated/triton.language.inline_asm_elementwise.html)