@DefTruth 2025-03-29
@LucasWilkinson closed this PR, PTAL for the new one https://github.com/vllm-project/vllm/pull/15742, I have add a workaround to fix this problem with mini overhead.